+++ packages/frontend/src/components/badge/NewItemBadge.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Badge, BadgeType } from '../Badge'
+import { Badge, BadgeType } from './Badge'
 
 export function NewItemBadge({ className }: { className?: string }) {
   return (
